change the constant and style
