--- packages/diva-app/src/constants/chainInfo.ts
@@ -1,58 +1,0 @@
-import { SupportedChainId } from './chains'
-import ethereumLogoUrl from '../Images/ethereum-logo.png'
-import polygonMaticLogo from '../Images/polygon-matic-logo.svg'
-import arbitrumLogoUrl from '../Images/arbitrum_logo.svg'
-
-interface BaseChainInfo {
-  readonly explorer: string
-  readonly logoUrl: string
-  readonly label: string
-  readonly nativeCurrency: {
-    name: string
-    symbol: string
-    decimals: number
-  }
-}
-
-export const CHAIN_INFO: { [chainId in SupportedChainId]: BaseChainInfo } = {
-  [SupportedChainId.MAINNET]: {
-    explorer: 'https://etherscan.io/',
-    label: 'Ethereum',
-    logoUrl: ethereumLogoUrl,
-    nativeCurrency: { name: 'Ether', symbol: 'ETH', decimals: 18 },
-  },
-  [SupportedChainId.ROPSTEN]: {
-    explorer: 'https://ropsten.etherscan.io/',
-    label: 'Ropsten',
-    logoUrl: ethereumLogoUrl,
-    nativeCurrency: { name: 'Ropsten Ether', symbol: 'ropETH', decimals: 18 },
-  },
-  [SupportedChainId.GOERLI]: {
-    explorer: 'https://goerli.etherscan.io/',
-    label: 'Görli',
-    logoUrl: ethereumLogoUrl,
-    nativeCurrency: { name: 'Görli Ether', symbol: 'görETH', decimals: 18 },
-  },
-  [SupportedChainId.POLYGON]: {
-    explorer: 'https://polygonscan.com/',
-    label: 'Polygon',
-    logoUrl: polygonMaticLogo,
-    nativeCurrency: { name: 'Polygon Matic', symbol: 'MATIC', decimals: 18 },
-  },
-  [SupportedChainId.POLYGON_MUMBAI]: {
-    explorer: 'https://mumbai.polygonscan.com/',
-    label: 'Polygon Mumbai',
-    logoUrl: polygonMaticLogo,
-    nativeCurrency: {
-      name: 'Polygon Mumbai Matic',
-      symbol: 'mMATIC',
-      decimals: 18,
-    },
-  },
-  [SupportedChainId.ARBITRUM_ONE]: {
-    explorer: 'https://arbiscan.io/',
-    label: 'Arbitrum',
-    logoUrl: arbitrumLogoUrl,
-    nativeCurrency: { name: 'Ether', symbol: 'ETH', decimals: 18 },
-  },
-}
