--- packages/diva-app/src/constants/chains.ts
@@ -1,17 +1,0 @@
-// List of all chains supported by the diva-app
-export enum SupportedChainId {
-  ROPSTEN = 3,
-  MAINNET = 1,
-  GOERLI = 5,
-  POLYGON = 137,
-  POLYGON_MUMBAI = 80001,
-  ARBITRUM_ONE = 42161,
-}
-
-// array of all chains id
-export const ALL_SUPPORTED_CHAIN_IDS: SupportedChainId[] = Object.values(
-  SupportedChainId
-).filter((id) => typeof id === 'number') as SupportedChainId[]
-
-// current supported chain
-export const CURRENT_SUPPORTED_CHAIN_ID = [SupportedChainId.ROPSTEN]
