--- packages/diva-app/src/constants/index.ts
@@ -1,2 +1,0 @@
-export * from './chainInfo'
-export * from './chains'
